Refactor about page
